Moved network client test helpers to their own files
